% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcfFilterPlotFunctions.R
\name{openGDSFile}
\alias{openGDSFile}
\title{Open a GDS file from a VCF file}
\usage{
openGDSFile(vcfFilePath, gdsFilePath)
}
\arguments{
\item{vcfFilePath}{Path to the VCF file.}

\item{gdsFilePath}{Path where the GDS file will be saved or is already saved.}
}
\value{
A GDS file object.
}
\description{
This function checks if a GDS file already exists for a given VCF file path.
If the GDS file does not exist, it converts the VCF file into GDS format
and then opens the GDS file.
}
\examples{
vcfFile <- system.file("extdata", "large_demo.vcf.gz", package = "vcfFilterPlot")
gdsFile <- system.file("extdata", "large_demo.vcf.gz", package = "vcfFilterPlot")
gdsObj <- openGDSFile(vcfFile, gdsFile)
}
