% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcfFilterPlotFunctions.R
\docType{data}
\name{ui}
\alias{ui}
\title{Shiny UI for VCF Missingness Analysis}
\format{
An object of class \code{shiny.tag.list} (inherits from \code{list}) of length 4.
}
\usage{
ui
}
\value{
A Shiny UI object.
}
\description{
This UI sets up the user interface for the VCF Missingness Analysis Shiny application.
It includes inputs for uploading a VCF file, setting thresholds for sample and site missingness, 
and for QUAL scores. It also provides action buttons to run the analysis.
}
\keyword{internal}
